+++ src/api-sdk/contracts/api-sdk-overview-profile.js
@@ -1,0 +1,61 @@
+export class ApiSdkOverviewProfile {
+  constructor({
+    profileName,
+    businessDrivers = [], objectives = [], architecturalSeparations = [], coreConcepts = [],
+    corePrinciples = [], productFields = [], consumerTypes = [], audienceClassifications = [],
+    interactionClassifications = [], logicalLayers = [], capabilities = [], contractFields = [],
+    authorityContextFields = [], errorClasses = [], lifecycleStates = [], sdkCapabilities = [],
+    sdkLayers = [], lifecycleStages = [], operatingRoles = [], qualityAttributes = [],
+    architecturalRules = [], architectureBoundaries = [], expectedOutcomes = [],
+    capabilityBeforeExposure = true, ownershipBeforePublication = true,
+    contractBeforeImplementation = true, identityBeforeProcessing = true,
+    authorityBeforeAction = true, validateTrustBoundaries = true,
+    compatibilityBeforeConvenience = true, idempotencyBeforeRetry = true,
+    errorsContractual = true, evidenceBeforeRetirement = true, providerOwnsDomain = true,
+    gatewayBounded = true, contractAuthoritative = true, derivedArtifactsTraceable = true,
+    productOwned = true, consumerRegistered = true, runtimeAuthorizationIndependent = true,
+    tenantPropertyEndToEnd = true, missingScopeFailsSafe = true,
+    asyncDeliverySeparateCompletion = true, errorsSafe = true, sideEffectsDefined = true,
+    breakingChangesMigrated = true, sdkSupportedProduct = true, sdkLayered = true,
+    sdkPreservesSemantics = true, docsVersionAligned = true, lifecycleGated = true,
+    agentAuthorityBounded = true, telemetryProtected = true, operationsReliable = true,
+    vendorNeutral = true, technologyIndependent = true,
+    apiMirrorsDatabase = false, gatewayOwnsBusinessLogic = false, reachabilityAuthorizes = false,
+    apiKeyCompleteIdentity = false, sdkOwnsDomainRules = false,
+    generatedArtifactAuthoritative = false, directDatabaseAccess = false,
+    oneTransportRequired = false, oneLanguageRequired = false, exactlyOnceGuaranteed = false,
+    sdkHidesProviderFailure = false, obsoletePreservedForever = false,
+    everyInternalCapabilityPublic = false, tenantSelfAssigned = false,
+    gatewayAcceptanceMeansCompletion = false, retryAllFailures = false,
+    agentDiscoveryExpandsAuthority = false, selectsApiProduct = false,
+    definesLaterChapterDetails = false
+  }) {
+    this.profileName = profileName;
+    for (const [key, value] of Object.entries({
+      businessDrivers, objectives, architecturalSeparations, coreConcepts, corePrinciples,
+      productFields, consumerTypes, audienceClassifications, interactionClassifications,
+      logicalLayers, capabilities, contractFields, authorityContextFields, errorClasses,
+      lifecycleStates, sdkCapabilities, sdkLayers, lifecycleStages, operatingRoles,
+      qualityAttributes, architecturalRules, architectureBoundaries, expectedOutcomes
+    })) this[key] = Object.freeze([...value]);
+    for (const [key, value] of Object.entries({
+      capabilityBeforeExposure, ownershipBeforePublication, contractBeforeImplementation,
+      identityBeforeProcessing, authorityBeforeAction, validateTrustBoundaries,
+      compatibilityBeforeConvenience, idempotencyBeforeRetry, errorsContractual,
+      evidenceBeforeRetirement, providerOwnsDomain, gatewayBounded, contractAuthoritative,
+      derivedArtifactsTraceable, productOwned, consumerRegistered,
+      runtimeAuthorizationIndependent, tenantPropertyEndToEnd, missingScopeFailsSafe,
+      asyncDeliverySeparateCompletion, errorsSafe, sideEffectsDefined, breakingChangesMigrated,
+      sdkSupportedProduct, sdkLayered, sdkPreservesSemantics, docsVersionAligned,
+      lifecycleGated, agentAuthorityBounded, telemetryProtected, operationsReliable,
+      vendorNeutral, technologyIndependent, apiMirrorsDatabase, gatewayOwnsBusinessLogic,
+      reachabilityAuthorizes, apiKeyCompleteIdentity, sdkOwnsDomainRules,
+      generatedArtifactAuthoritative, directDatabaseAccess, oneTransportRequired,
+      oneLanguageRequired, exactlyOnceGuaranteed, sdkHidesProviderFailure,
+      obsoletePreservedForever, everyInternalCapabilityPublic, tenantSelfAssigned,
+      gatewayAcceptanceMeansCompletion, retryAllFailures, agentDiscoveryExpandsAuthority,
+      selectsApiProduct, definesLaterChapterDetails
+    })) this[key] = Boolean(value);
+    Object.freeze(this);
+  }
+}